requirements.txt
PyGithub==1.59.1
python-dotenv==1.0.1
qdrant-client==1.9.2
openai>=1.0.0
